fix: properly handle double-quoted strings within string interpolations
Closes https://github.com/decaffeinate/decaffeinate-parser/issues/14.
Closes https://github.com/decaffeinate/decaffeinate/issues/229.
Closes https://github.com/decaffeinate/decaffeinate/issues/435.

The code that builds interpolation tokens had a case where it handled
intermediate string fragments in the template string, but sometimes it would
falsely identify a double-quoted string inside an interpolation as one of these
string fragments. Fortunately, we already track a boolean for whether we're
inside an interpolation or not, so we can just add that boolean to the check.

This issue ended up causing the range of the string template to be interpreted
incorrectly (the end would be too soon), which caused an infinite loop in
decaffeinate-parser when it searched backward from the end until the first '}'
character.

In addition to running the coffee-lex tests, I also ran the decaffeinate-parser
and decaffeinate tests with this patch applied and they all pass.