fix: allow newlines when looking for identifiers after DOT tokens (#52)
Closes https://github.com/decaffeinate/decaffeinate/issues/437.

If we see a DOT token, then the next token shouldn't be treated as a keyword
because it should instead be seen as a property access. This was already handled
in the code, but it's even true when there's a newline in between the dot and
the keyword. So, when computing the previous token, we want to skip any NEWLINE
tokens we see, and check if the next earlier one is a DOT (or PROTO or AT).