refactor: rework some string lexing to avoid stringLocationsFromString
Progress toward https://github.com/decaffeinate/decaffeinate/issues/557

Previously, if you had a string like `"a#{b}c"`, it would lex like this:

DSTRING, INTERPOLATION_START, IDENTIFIER, INTERPOLATION_END, DSTRING

then the `stringLocationsFromString` function would do another pass and change
it to use `STRING_START`, `STRING_CONTENT`, and `STRING_END`. It would detect
the end of a string by finding a `DSTRING` not followed by an
`INTERPOLATION_START`.

Now, instead of that separate pass, we generate `STRING_START`, `STRING_END`,
and `STRING_CONTENT` nodes during the main lexing pass (for single and double
quoted strings) by tracking the type of string on a new `stringStack`. One nice
thing about this change is that there is no longer a distinction between strings
with interpolations and strings without interpolations; they always start with a
`STRING_START` and end with a `STRING_END`.

The plan for future commits is to change herestrings and heregexes to use this
same approach, and then to add a follow-up pass that changes the
`STRING_CONTENT` tokens into something more granular that expresses which parts
of the string should be trimmed.

There's a reasonable chance this would be a breaking change to external tools,
but it doesn't seem to cause any problems in decaffeinate-parser or decaffeinate
(the tests all pass, and they never used the SSTRING and DSTRING tokens that I'm
removing). So my plan is for this to just be a refactor commit without a
release, and probably in the near future there will be a real breaking release
with a major version bump.