feat: calculate padding on single and double quoted strings (#68)
Progress toward these issues:
https://github.com/decaffeinate/decaffeinate/issues/557
https://github.com/decaffeinate/decaffeinate/issues/554
https://github.com/decaffeinate/decaffeinate/issues/406,

This change inserts STRING_PADDING and STRING_LINE_SEPARATOR tokens into single
and double quoted strings such that the proper string contents can be attained
by removing STRING_PADDING ranges and changing STRING_LINE_SEPARATOR ranges to
spaces. This will be particularly useful when using magic-string in
decaffeinate.

To make this easier, I created a PaddingTracker class that abstracts away the
need to deal with ranges. Instead, it provides the contents of each string
fragment and lets you call `markPadding` and `markLineSeparator` to designate
ranges as padding or a line separator, and then figures out the details for
transforming the old to new tokens.

The plan is to use this same class when computing heregexp padding, and I'll
also probably try to change the `tripleQuotedStringSourceLocations` function to
also use `PaddingTracker`.

As part of testing this, I set up some test helpers for comparing the coffee-lex
output to the output from the CoffeeScript lexer, so that I can make sure that
all test cases match CoffeeScript's behavior.