refactor: update tests and name for triple quoted string function (#71)
Progress toward these issues:
https://github.com/decaffeinate/decaffeinate/issues/557
https://github.com/decaffeinate/decaffeinate/issues/556

The plan is to make it so all three string padding functions are structured and
tested similarly, and this commit moves the code closer to that.

In addition to changing the tests to use `verifyStringMatchesCoffeeScript`, I
added a bunch of tests from parseLiteral_test.js in decaffeinate-parser. Some of
these tests are failing due to bugs in the coffee-lex implementation, so I
marked those as skipped and will fix those in a follow-up commit.