fix: ensure multiple copies of coffee-lex can coexist
Because both decaffeinate-parser and decaffeinate depend on coffee-lex, it's possible for their versions to be mismatched such that npm will install multiple copies of coffee-lex. This results in problems because the tokens coffee-lex generates have a source type that is an instance of the `SourceType` class, meaning that the same source type in different copies will not be equal to each other, making checking that a token is of a particular type, e.g. `CALL_START`, impossible if the token was created using a different version of coffee-lex than the one used to check the type.

  BREAKING CHANGE: To fix this I've changed the `SourceType` class to be an `enum`, so its values will just be explicit numbers that will compare correctly across versions of coffee-lex. Because this changes the API of the "instances" of `SourceType` by removing the `name` property, this is a breaking change. If the `name` of a source type is desired, it can be looked up on the enum type: `SourceType[SourceType.AT] === 'AT'`.