fix: lex a number with a leading dot as one token (#90)
Previously, this would be lexed as a `DOT` followed by a `NUMBER`. The regular expression we use to match actually does work with numbers with leading dots, but we were consuming `.` by itself with higher precedence. To fix it, we just move the `NUMBER` match up ahead of the `DOT`.