fix: properly calculate herestring indentation before interpolations (#97)
Fixes https://github.com/decaffeinate/decaffeinate/issues/927

Our test for "is this a whitespace-only line" didn't quite work, since it
returned true for the end of a fragment that was followed by an interpolation.
Now, we limit that test to cases where the line we have is actually the full
line from the source code, i.e. intermediate lines in fragments and the last
line of the last fragment.