fix: `SourceTokenList#[Symbol.iterator]` API incorrect (#289)
This fix was flagged by an upgrade to TypeScript and eslint, which in the process changed a bunch of other things due to plugin/config updates.

For some reason the `lint` script was not actually running on all the files. This fixes it by letting eslint itself handle the globbing rather than the shell.